move history in section
