+++ scripts/scripts.js
@@ -14,6 +14,11 @@ import {
   setLanguage, createMetadata,
 } from './lib-franklin.js';
 
+import {
+  decorateHistorySection,
+  observeHistorySection,
+} from './lib-history-section.js';
+
 const LCP_BLOCKS = []; // add your LCP blocks to the list
 window.hlx.RUM_GENERATION = 'mammotome'; // add your RUM generation information here
 
@@ -45,94 +50,6 @@ function buildAutoBlocks(main) {
   }
 }
 
-function getNextSiblings(child, parent, until) {
-  let isNextSiblings = false;
-  const nextSiblings = [];
-  parent.childNodes.forEach((node) => {
-    if (until === 'UNTIL_NEXT_H3' && node.tagName.toLowerCase() === 'h3') {
-      isNextSiblings = false;
-    }
-    if (isNextSiblings) {
-      nextSiblings.push(node);
-    }
-    if (node === child) {
-      nextSiblings.push(node);
-      isNextSiblings = true;
-    }
-  });
-  return nextSiblings;
-}
-
-/**
- * Decorates the history page.
- * @param {Element} main The main element
- */
-// eslint-disable-next-line import/prefer-default-export
-export async function decorateHistoryPage(main) {
-  try {
-    const parent = main.querySelector('.default-content-wrapper');
-    const firstH3 = parent.querySelector('h3');
-
-    // wrap all h3 and p in a timeline element
-    const timelineBlock = buildBlock('timeline', null);
-    parent.insertBefore(timelineBlock, firstH3);
-    timelineBlock.append(...getNextSiblings(firstH3, parent));
-
-    // create small blocks for every year
-    const yearBlocks = [];
-    timelineBlock.childNodes.forEach((el) => {
-      if (el.tagName.toLowerCase() === 'h3') {
-        const yearBlock = buildBlock('year', null);
-        yearBlocks.push([yearBlock, getNextSiblings(el, timelineBlock, 'UNTIL_NEXT_H3')]);
-      }
-    });
-    yearBlocks.forEach((block) => {
-      timelineBlock.append(block[0]);
-      block[0].append(...block[1]);
-    });
-
-    // change order and move picture element before h3 element. needed for mobile.
-    yearBlocks.forEach((block) => {
-      block[0].childNodes.forEach((el) => {
-        if (el.tagName.toLowerCase() === 'h3' && el.nextSibling?.firstElementChild?.tagName?.toLowerCase() === 'picture') {
-          block[0].insertBefore(el.nextSibling, el);
-        }
-      });
-    });
-  } catch (error) {
-    // eslint-disable-next-line no-console
-    console.error('Decorating History page failed', error);
-  }
-}
-/**
- * Observe the history page for changes.
- * @param {Element} main The main element
-*/
-// eslint-disable-next-line import/prefer-default-export
-export async function observeHistoryPage(main) {
-  try {
-    const highlightWhenInViewport = new IntersectionObserver((entries) => {
-      entries.forEach((entry) => {
-        const { isIntersecting } = entry;
-        if (isIntersecting) {
-          entry.target.classList.add('highlight');
-        } else {
-          entry.target.classList.remove('highlight');
-        }
-      });
-    }, {
-      rootMargin: '0px',
-      threshold: 1,
-    });
-
-    const yearBlocks = main.querySelectorAll('.timeline .year');
-    yearBlocks.forEach((el) => highlightWhenInViewport.observe(el));
-  } catch (error) {
-    // eslint-disable-next-line no-console
-    console.error('Observer of History page failed', error);
-  }
-}
-
 /**
  * Decorates the main element.
  * @param {Element} main The main element
@@ -145,8 +62,11 @@ export async function decorateMain(main) {
   buildAutoBlocks(main);
   decorateSections(main);
   decorateBlocks(main);
-  decorateHistoryPage(main);
-  observeHistoryPage(main);
+
+  if (main.querySelector('.section.our-history')) {
+    decorateHistorySection(main);
+    observeHistorySection(main);
+  }
 }
 
 /**
